Update paths to be relative to the generated file.
