+++ Weave/GenerateWeaveSources.cs
@@ -2,7 +2,6 @@
 
 namespace Weave
 {
-    using System;
     using System.CodeDom.Compiler;
     using System.IO;
     using System.Linq;
@@ -156,21 +155,6 @@ namespace Weave
                 isEnabledByDefault: true,
                 warningLevel: level,
                 location: Location.Create(error.FileName, new TextSpan(0, 0), new LinePositionSpan(linePosition, linePosition)));
-        }
-
-        private static class PathUtils
-        {
-            public static string MakeRelative(string root, string path)
-            {
-                root = Path.GetFullPath(root);
-                var fullFilePath = Path.GetFullPath(Path.Combine(root, path));
-                var relativeUri = new Uri(root).MakeRelativeUri(new Uri(fullFilePath));
-                var relativePath = Uri.UnescapeDataString(relativeUri.ToString()).Replace(Path.AltDirectorySeparatorChar, Path.DirectorySeparatorChar);
-                return relativePath;
-            }
-
-            public static string EncodePathAsFilename(string path) =>
-                Uri.EscapeDataString(path).Replace("_", "__").Replace("%", "_p");
         }
     }
 }
